Implement Clone feature

+++ domain/monster-lore/data/src/commonMain/kotlin/br/alexandregpereira/hunter/data/monster/lore/local/MonsterLoreLocalDataSource.kt
@@ -42,9 +42,10 @@ internal class MonsterLoreLocalDataSource(
 
     fun save(
         monsters: List<MonsterLoreCompleteEntity>,
+        isSync: Boolean = false
     ): Flow<Unit> = flow {
         mutex.withLock {
-            deleteAllData()
+            if (isSync) deleteAllData()
             insertAll(monsters)
         }
         emit(Unit)
